feat(opencode): rewrite sync-pull/push to take { directory } and target IssueTable
ADR-0003 step 4. The previous sync layer in src/linear/ predated the
new IssueTable (introduced in step 1) and imported Session.Todo.Service
to manage its rows. It also returned a misleading "Already up to date"
label when Linear had issues that weren't yet linked locally — the
exact bug raised in the open issue ("Pull from Linear hints 'Already up
to date (1 already synced)' but Todos shows no todo").

SyncPull.pull({ directory }) now:
  - fetches active Linear issues (unstarted/started) for the configured
    project (paginated, 50/page)
  - inserts any whose `linear_issue_id` is not already linked locally
  - NEVER overwrites existing local rows (local edits are first-class
    per ADR-0002 D5)
  - returns honest pulled/skipped/failed counts — no "already up to
    date" euphemism (ADR-0002 D6)

SyncPush.push({ directory, issueIds? }) now:
  - targets IssueTable (not Session.Todo.Service)
  - skips issues with no `linear_issue_id` (local-only)
  - skips issues whose `time_updated <= last_pushed_at`
  - on success, sets `last_pushed_at = time_updated` in a single SQL
    UPDATE, keeping the watermark in lockstep so the next push skips
    unchanged rows (Drizzle's `$onUpdate` would otherwise bump
    `time_updated` past `last_pushed_at` and force a re-push)
  - new issues are NOT created by bulk push; that's an explicit
    per-row action deferred to a future ADR

Both modules now consume the Linear MCP client via a Context tag
(`SyncPull.Client` / `SyncPush.Client`) and read config from
`Config.Service`, instead of pulling module-level state.

Tests: 5/5 pass. The integration test in src/linear/ is deleted — it
targeted the old Session.Todo.Service and is superseded by the focused
unit tests under src/issue/.

The shims in src/linear/ now re-export from src/issue/, so any
existing imports of `linear/sync-pull` and `linear/sync-push` keep
working. (These shims go away in step 10.)
